Fix crash if a double value returned for the 'rating' field in the video API response

+++ Application/ZypeAPI/src/main/java/com/zype/fire/api/Model/VideoData.java
@@ -68,7 +68,7 @@ public class VideoData {
     public String publishedAt;
 
     @Expose
-    public int rating;
+    public double rating;
 
     @SerializedName("related_playlist_ids")
     @Expose
